requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1
pydantic[email]==2.5.0
pydantic-settings==2.1.0
google-cloud-aiplatform==1.38.1
firebase-admin==6.2.0
redis==5.0.1
httpx==0.25.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
celery==5.3.4
prometheus-fastapi-instrumentator==6.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0